requirements.txt
Django==2.2
django-environ==0.4.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
PyJWT==2.0.1
python-dotenv==0.17.0
pytz==2021.1
sqlparse==0.4.1